refactor: update according to PR comment and factorize 'verify_digest'
