+++ src/core/actix/mod.rs
@@ -1,2 +1,1 @@
 pub mod inbox;
-mod signature;
