+++ src/core/axum/mod.rs
@@ -11,7 +11,6 @@ use digest::{verify_sha256, DigestPart};
 mod digest;
 pub mod inbox;
 pub mod json;
-mod signature;
 
 /// A request guard to ensure digest has been verified request has been
 /// see [`receive_activity`]
